requirements.txt
click~=8.0.4
nbformat~=5.7.3
fastapi
uvicorn~=0.23.1
pydantic
numpy
autopep8~=2.0.2
python-dotenv
termcolor~=2.3.0
websockets
APScheduler
Jinja2~=3.0.3
PyYAML~=6.0
setuptools~=68.0.0
requests~=2.31.0